The columns for viewing raw data can now be persisted after the program is closed through a file called columns.cfg. The syntax for columns.cfg is follows:
shown|expression_for_column1|alias_for_column1

where shown can be 0, for not shown, and 1, for shown. 

Such a file can be created either manually or from the save button in the
Choose Column dialog, and if such a file exists, it would be read; if not, the
default (all branches from branches.txt) will be used.

----------

the remove column function is still not working! still trying...